Exported an `encrypt` function from `cipher.js`
`encrypt()` takes in two arguments, an alphanumeric input to encrypt, and a integer representing the
number of characters to shift when encrypting.

`encrypt()` uses an array of the alphabet plus numbers 0 through 9, and `reduce()`s that
array into a map where each letter/number is a key pointing to its encrypted (shifted) counterpart.
For example one key value pair, if `shift` is `3`, would be `{ a: 'd' }`.

`encrypt()` then iterates over each character of the input, adding the encrypted version of that
character to the accumulated result of the call to `reduce()`.