Ensure input is a string, handle non-alphanumerics
Added a couple functions to `cipher.js` for use in `encrypt()`. The first checks that the input
passed to `encrypt()` is a string, the second checks that a given string is an alphanumeric
character.

If the input passed to `encrypt()` is not a string, a blank string is returned. `encrypt()` should
always return a string in case it is being used as part of a larger system that relies upon
its return value being a string.

If a character in the string input is non-alphanumeric `encrypt()` does not encrypt that character,
and instead adds it as-is to the encrypted string.

Unit tests added for these scenarios.